updating int for long

+++ crud/TesteDotnet.Domain/Entities/Contato.cs
@@ -5,9 +5,9 @@ public class Contato
     public Guid PessoaId { get; set; }
 
     public string Nome { get; set; }
-    public int Celular { get; set; }
+    public long Celular { get; set; }
 
-    public Contato(Guid pessoaId, string nome, int celular)
+    public Contato(Guid pessoaId, string nome, long celular)
     {
         PessoaId = pessoaId;
         Nome = nome;
