+++ crud/TesteDotnet.Domain/Entities/Pessoa.cs
@@ -7,10 +7,10 @@ public class Pessoa
     public string Nome { get; }
     public string Email { get; }
     
-    public int CPF { get; }
+    public long CPF { get; }
 
     // Info that may be null
-    public int? CEP { get; set; }
+    public long? CEP { get; set; }
     public string? Logradouro { get; set; } 
     public string? Complemento { get; set; }
     public string? Bairro { get; set; }
@@ -18,7 +18,7 @@ public class Pessoa
 
     public List<Contato> Contatos { get; } = new List<Contato>();
 
-    public Pessoa(string nome, string email, int cpf)
+    public Pessoa(string nome, string email, long cpf)
     {
         CPF = cpf;
         Nome = nome;
