+++ crud/TesteDotnet.Domain/Interfaces/Repositories/IContatoRepository.cs
@@ -2,7 +2,7 @@ using TesteDotnet.Domain.Entities;
 
 namespace TesteDotnet.Domain.Interfaces.Repositories;
 
-public interface IContatoRepository : IBaseRepository<int, Contato>
+public interface IContatoRepository : IBaseRepository<long, Contato>
 {
     
 }
